Rearrange so gem works.

+++ ext/swf_render/flash_rasterizer.h
@@ -4,6 +4,6 @@
 int render(char* input_swf, char* output_png);
 int render_to_buffer(char* input_swf,
                      unsigned char** out,
-                     size_t* outsize);
+                     unsigned int* outsize);
 
 #endif
